+++ ui/src/components/DocIdDisplay.vue
